requirements.txt
opencv-contrib-python>=4.7
mediapipe>=0.9
numpy>=1.26
scipy>=1.11